Changed simulate.py to compare particle with simulation
Changed simulate.py so it simulates the experimental setup.  Changed bug in tracking.py that gave 8 times higher diffusion constants (particle raddii seems unaffected) and last few frames of frames can be dropped so that the wrong background subtraction for the last 20 frames does not affect the obtained diffusion constants and particle diameters.