Added more selection parameters for autoCropper.py
Added a image crop for the whole image: all particles with a distance from the boundary can be ignored. If there is too much scattering from the electrodes, the image can be made smaller. A minimum mass can be selected and the bug that caused that the original image was used for selecting features except a blurred one is solved. Also the list of parameters can now be fed to the script from a file.